backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
chromadb
sentence-transformers
uuid
python-docx
python-pptx
pdf2image
pytesseract
Pillow
